fix(alerts): better nerdgraph operator usage
Without this change, the nerdgraph operator for nrql conditions is not accurate.
The nerdgraph type uses the term "EQUALS" in place of the older rest term
"EQUAL".  This was never caught, because we were not using that particular
operator on any of our tests.  Here we include the
AlertsNrqlConditionTermsOperator type to be generated, and move references to it
with a small amount of test coverage for this behavior.