fix(http): move logger initialization to NewClient()
Without this change, the logger may or may not be established by the time a call
is made.  This can cause conflict in the case of a global logger.  Here we
attach a logger field to the Client and ensure that it is set up during the
NewClient call.  Also updates usage within the Client to use the attached
logger.