fix: omit filter ID when empty
This is a fix for https://github.com/newrelic/terraform-provider-newrelic/issues/1986 . When trying to update a workflow via terraform an error is thrown due to the fact that no filter ID (empty) is passed the first time. By allowing this to be omitted when empty (which is still a valid request) the terraform update of a workflow works and results in a valid workflow.