fix(workflows)!: forward empty destinations correctly on workflows update
BREAKING CHANGE: this commit changes type of the DestinationConfigurations field of AiWorkflowsUpdateWorkflowInput, breaking previous usages

While workflow must have at least one destination, the current behaviour is very confusing: if the list is empty, the client will silently ignore it.
The new behaviour would simply pass the empty list to the backend which will determine how to handle it (in this case, return a validation error).