fix(workflows)!: forward empty workflow name update correctly
BREAKING CHANGE: this commit changes type of the Name field of AiWorkflowsUpdateWorkflowInput, breaking previous usages
While empty string is not a valid name, the current behaviour is very confusing: if the name is an empty string, the client will silently ignore it.
The new behaviour would simply pass the empty string to the backend which will determine how to handle it (in this case, return a validation error).